refactor: enhance server setup by separating message service initialization and improving process handlers
