+++ src/config/server/index.ts
@@ -1,10 +1,12 @@
-import express from 'express';
+import express, { Application } from 'express';
 import path from 'path';
 import { errorHandler } from '../../middleware/error/errorHandler';
-import { requestLogger } from '../../middleware/logger';
+import { requestLogger } from '../../middleware/requestLogger';
 import { MessageDb } from '../../database/messageDb';
 import { MessageService } from '../../services/messageService';
 import { createMessageRoutes } from '../../routes/messageRoutes';
+
+let messageService: MessageService;
 
 /**
  * Creates and configures an Express application.
@@ -15,29 +17,40 @@ import { createMessageRoutes } from '../../routes/messageRoutes';
  * - Uses a custom request logger middleware.
  * - Uses a custom error handler middleware (should be the last middleware).
  *
- * @returns {express.Application, MessageService} The configured Express application instance.
+ * @returns {Application} The configured Express application instance.
  */
-export function createApp(): [express.Application, MessageService] {
+export function createApp(): Application {
   const app = express();
 
   // Static files
-  const publicDir = path.join(__dirname, '../../../public');
-  app.use(express.static(publicDir));
+  app.use(express.static(path.join(__dirname, '../public')));
 
   // Middleware
   app.use(express.json());
   app.use(requestLogger);
+
   // Initialize services
   const messageDb = new MessageDb();
-  const messageService = new MessageService(messageDb);
+  messageService = new MessageService(messageDb);
 
   // Routes
-  app.get('/', (_req, res) => {
-    res.sendFile(path.join(publicDir, 'index.html'));
-  });
   app.use('/api/messages', createMessageRoutes(messageService));
+
   // Error handling - should be last
   app.use(errorHandler);
 
-  return [app, messageService];
+  return app;
+}
+
+/**
+ * Gets the initialized message service.
+ *
+ * @returns {MessageService} The message service instance.
+ * @throws {Error} If the message service has not been initialized.
+ */
+export function getMessageService(): MessageService {
+  if (!messageService) {
+    throw new Error('MessageService has not been initialized');
+  }
+  return messageService;
 }
