+++ src/utils/server/processHandlers.ts
@@ -1,5 +1,19 @@
 import Logger from '../../logger';
 
+/**
+ * Sets up process handlers for handling unhandled promise rejections,
+ * uncaught exceptions, and SIGTERM signals.
+ *
+ * This function performs the following steps:
+ * 1. Listens for unhandled promise rejections and logs the error before exiting the process.
+ * 2. Listens for uncaught exceptions and logs the error before exiting the process.
+ * 3. Listens for SIGTERM signals (e.g., from `kill` command or Docker stop) and logs a message before exiting the process.
+ *
+ * @example
+ * ```typescript
+ * setupProcessHandlers();
+ * ```
+ */
 export function setupProcessHandlers(): void {
   process.on('unhandledRejection', (error: Error) => {
     Logger.error('Unhandled Promise Rejection:', error);
@@ -12,12 +26,7 @@ export function setupProcessHandlers(): void {
   });
 
   process.on('SIGTERM', () => {
-    Logger.info('SIGTERM received. Performing graceful shutdown...');
-    process.exit(0);
-  });
-
-  process.on('SIGINT', () => {
-    Logger.info('SIGINT received. Performing graceful shutdown...');
+    Logger.info('Received SIGTERM, shutting down gracefully');
     process.exit(0);
   });
 }
